Small fixes
I have fixed line 4 which was missing some dependencies. I have also
removed the further steps after compiling Bitcoin as some users may not
have a memory stick plugged in and ready. That should be in a separate
script I think which we could eventually call from this one with an
input prompt.

Also Line 6 was returning an error, it was just a minor fix to select
line 16 just do ‘16s/.*/…/g’
The -i flag just means edit in place so it can all be done in one line.

Later on in the script I remove the swap file to save space. However it
did work with only 100Mb.

I have also started to add comments to the code so that people can see
what it’s doing. Please feel free to add more details such as the
Berkely DB, installing Bitcoin, including telling the user which branch.

This is a good effort and hopefully you will get more people in to
coding as now they will see for themselves the benefits of automating
otherwise laborious tasks.